requirements.txt
instapy==0.6.16
PySocks==1.7.1
python-dotenv==1.0.0
emoji==1.2.0
clarifai==2.6.2
selenium==4.9.0